Replaced near-cube snapping with grid snapping. Moved Cube to cube.py
